+++ deepcave/__init__.py
@@ -1,17 +1,29 @@
 import sys
+import datetime
 
-author = "René Sass and  Marius Lindauer"
-version = "0.2"
-
+name = "DeepCAVE"
+package_name = "deepcave"
+author = "René Sass and Marius Lindauer"
+author_email = "sass@tnt.uni-hannover.de"
+description = "An interactive framework to visualize and analyze your AutoML process in real-time."
+url = "automl.org"
+project_urls = {
+    "Documentation": "https://automl.github.io/DeepCAVE/main",
+    "Source Code": "https://github.com/automl/deepcave",
+}
+copyright = f"Copyright {datetime.date.today().strftime('%Y')}, René Sass and Marius Lindauer, <<author_email>>"
+version = "0.3"
 
 _exec_file = sys.argv[0]
+_exec_files = ["server.py", "worker.py", "sphinx-build"]
 
-if "server.py" in _exec_file or "worker.py" in _exec_file:
+
+if any(file in _exec_file for file in _exec_files):
     from deepcave.config import config
     from deepcave.queue import Queue  # noqa
     from deepcave.runs.handler import RunHandler  # noqa
-    from deepcave.runs.objective import Objective
-    from deepcave.runs.recorder import Recorder
+    from deepcave.runs.objective import Objective  # noqa
+    from deepcave.runs.recorder import Recorder  # noqa
     from deepcave.server import get_app  # noqa
     from deepcave.utils.cache import Cache  # noqa
     from deepcave.utils.run_caches import RunCaches  # noqa