+++ deepcave/utils/dash.py
@@ -3,19 +3,9 @@ import dash_bootstrap_components as dbc
 
 def flash(message: str, category: str = "info"):
     """
-    Flask style flash-message with Alerts
-    https://dash-bootstrap-components.opensource.faculty.ai/docs/components/alert/
-
-    Possible categories:
-        primary    -> blue
-        secondary  -> light grey
-        success    -> green
-        warning    -> yellow
-        danger     -> red
-        info       -> cyan
-        light      -> white
-        dark       -> dark gray
+    Flask style flash-message with Alerts.
     """
+
     return dbc.Alert(
         message,
         id=f"alert_{hash(message)}",